test(bdd): declare encrypted cache backing
